client: merge convertable functionality into currency
